bones_scripting

Module lua

Re-exports

pub use piccolo;

Modules

bindings

Structs

LuaEngine

Resource used to access the lua scripting engine.

LuaPlugin

A lua plugin asset.

LuaPluginLoaderSessionPlugin

A [`SessionPlugin] that will run the provided lua plugins

LuaPluginSystems

The systems that have been registered by a lua plugin.

LuaPlugins

Resource containing the lua plugins that have been installed in this session.

LuaScript

A Lua script asset.

LuaSingletons

Struct for accessing and initializing lua singletons.

WorldRef

A frozen reference to the ECS [World].

Enums

LuaPluginSystemsState

The load state of the [`LuaPluginSystems

Traits

CtxExt

Extension trait for the [Context] that makes it easier to access our lua singletons.

MetatableFn

Helper trait to get a singleton fn pointer for the metatable for a type.

ValueExt

Extension trait on top of [Value] to add helper functions.

Functions

lua_game_plugin

Install the lua scripting plugin.

Type Aliases

LuaPluginSystemsCell

An atomic cell containing the LuaPluginSystemsState.

SystemStageId

The ID of a system stage.