Module bones_scripting::lua::bindings::ecsref

Structs

AssetRef

A reference to an asset in an EcsRef

ComponentBorrow

A component borrow into an EcsRef.

ComponentBorrowMut

A mutable component borrow into an EcsRef.

ComponentRef

A reference to component in an EcsRef.

EcsRef

A reference to an ECS-compatible value.

EcsRefBorrow

A borrow of an EcsRef.

EcsRefBorrowMut

A mutable borrow of an EcsRef.

SchemaLuaEcsRefMetatable

A type data that can be used to specify a custom metatable to use for the type when it is used in an EcsRef in the lua API.

Enums

EcsRefBorrowError

An error that occurs when borrowing an EcsRef.

EcsRefBorrowKind

A kind of borrow into an EcsRef.

EcsRefBorrowMutKind

A kind of mutable borrow of an EcsRef.

EcsRefData

The kind of value reference for EcsRef.

Functions

metatable